pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepaper"
version = "0.1.0"
description = "Paper knowledge base skill for Claude Code. Analyze arxiv papers with /deepaper."
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["arxiv", "paper", "research", "claude", "obsidian", "ai", "academic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "typer>=0.9",
    "httpx>=0.25",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/lzw19951010/deepaper"
Repository = "https://github.com/lzw19951010/deepaper"
Issues = "https://github.com/lzw19951010/deepaper/issues"

[project.scripts]
deepaper = "deepaper.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
